backend/requirements.txt
# ── Prompt 1 core dependencies ────────────────────────────────────────────────
openai>=1.30.0
pyairtable>=2.3.0
pydantic>=2.6.0
jsonschema>=4.22.0
python-docx>=1.1.0
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
redis>=5.0.0
tenacity>=8.3.0

# ── Prompt 2 — API / Auth / Queue ─────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
starlette>=0.37.0

# OAuth
authlib>=1.3.0
httpx>=0.27.0          # Authlib async HTTP transport

# Session cookie signing
itsdangerous>=2.2.0

# Queue
celery[redis]>=5.4.0

# SQLite type handling (datetime etc.)
# stdlib sqlite3 is sufficient — no extra package needed

# File parsing (already partially in P1, confirmed here)
# python-docx — already above
# pdfplumber  — already above
